requirements.txt
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2
pymongo==4.6.0
flask==3.0.0
numpy==1.26.0
seaborn==0.13.0
matplotlib==3.8.0
jupyter==1.0.0
notebook==7.0.0
flask==2.3.3
kafka-python==2.2.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0